backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.1
geopy==2.4.1
redis==5.0.1